Backend/requirements.txt
--index-url https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/cu121

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Deep Learning & Computer Vision (GPU-enabled)
torch==2.9.0
torchvision==0.24.0
torchaudio==2.9.0
ultralytics==8.3.221
opencv-python==4.12.0.88
cvzone==1.6.1

# Scientific Computing & Image Processing
numpy==2.2.6
scipy==1.16.2
scikit-image==0.25.2
pillow==11.3.0
imageio==2.37.0

# Data Processing
polars==1.34.0
pandas 

# Visualization
matplotlib==3.10.7

# Jupyter & Development Tools
ipykernel==7.0.1
ipython==9.6.0
jupyter_client==8.6.3

# Utilities
PyYAML==6.0.3
requests==2.32.5
tqdm